Rename: webgpu/builder to webgpu/nodes
