+++ examples/jsm/renderers/webgpu/nodes/WGSLNodeFunction.js
